Bring in and fix the contact module tests
