+++ app-modules/contact/tests/Tenant/Contact/RequestFactories/EditContactRequestFactory.php
@@ -1,0 +1,70 @@
+<?php
+
+/*
+<COPYRIGHT>
+
+    Copyright © 2016-2025, Canyon GBS LLC. All rights reserved.
+
+    Aiding App™ is licensed under the Elastic License 2.0. For more details,
+    see <https://github.com/canyongbs/aidingapp/blob/main/LICENSE.>
+
+    Notice:
+
+    - You may not provide the software to third parties as a hosted or managed
+      service, where the service provides users with access to any substantial set of
+      the features or functionality of the software.
+    - You may not move, change, disable, or circumvent the license key functionality
+      in the software, and you may not remove or obscure any functionality in the
+      software that is protected by the license key.
+    - You may not alter, remove, or obscure any licensing, copyright, or other notices
+      of the licensor in the software. Any use of the licensor’s trademarks is subject
+      to applicable law.
+    - Canyon GBS LLC respects the intellectual property rights of others and expects the
+      same in return. Canyon GBS™ and Aiding App™ are registered trademarks of
+      Canyon GBS LLC, and we are committed to enforcing and protecting our trademarks
+      vigorously.
+    - The software solution, including services, infrastructure, and code, is offered as a
+      Software as a Service (SaaS) by Canyon GBS LLC.
+    - Use of this software implies agreement to the license terms and conditions as stated
+      in the Elastic License 2.0.
+
+    For more information or inquiries please visit our website at
+    <https://www.canyongbs.com> or contact us via email at legal@canyongbs.com.
+
+</COPYRIGHT>
+*/
+
+namespace AidingApp\Contact\Tests\Tenant\Contact\RequestFactories;
+
+use AidingApp\Contact\Models\ContactSource;
+use AidingApp\Contact\Models\ContactStatus;
+use App\Models\User;
+use Worksome\RequestFactories\RequestFactory;
+
+class EditContactRequestFactory extends RequestFactory
+{
+    public function definition(): array
+    {
+        $firstName = $this->faker->firstName();
+        $lastName = $this->faker->lastName();
+
+        return [
+            'status_id' => ContactStatus::inRandomOrder()->first()?->id ?? ContactStatus::factory()->create()->id,
+            'source_id' => ContactSource::inRandomOrder()->first()?->id ?? ContactSource::factory()->create()->id,
+            'first_name' => $firstName,
+            'last_name' => $lastName,
+            'full_name' => "{$firstName} {$lastName}",
+            'preferred' => $this->faker->firstName(),
+            'description' => $this->faker->paragraph(),
+            'email' => $this->faker->email(),
+            'mobile' => $this->faker->e164PhoneNumber(),
+            'sms_opt_out' => $this->faker->boolean(),
+            'email_bounce' => $this->faker->boolean(),
+            'phone' => $this->faker->e164PhoneNumber(),
+            'address' => $this->faker->address(),
+            'address_2' => $this->faker->address(),
+            'assigned_to_id' => User::factory()->create()->id,
+            'created_by_id' => User::factory()->create()->id,
+        ];
+    }
+}
