+++ app-modules/contact/tests/Tenant/Organization/RequestFactories/CreateOrganizationRequestFactory.php
@@ -1,0 +1,67 @@
+<?php
+
+/*
+<COPYRIGHT>
+
+    Copyright © 2016-2025, Canyon GBS LLC. All rights reserved.
+
+    Aiding App™ is licensed under the Elastic License 2.0. For more details,
+    see <https://github.com/canyongbs/aidingapp/blob/main/LICENSE.>
+
+    Notice:
+
+    - You may not provide the software to third parties as a hosted or managed
+      service, where the service provides users with access to any substantial set of
+      the features or functionality of the software.
+    - You may not move, change, disable, or circumvent the license key functionality
+      in the software, and you may not remove or obscure any functionality in the
+      software that is protected by the license key.
+    - You may not alter, remove, or obscure any licensing, copyright, or other notices
+      of the licensor in the software. Any use of the licensor’s trademarks is subject
+      to applicable law.
+    - Canyon GBS LLC respects the intellectual property rights of others and expects the
+      same in return. Canyon GBS™ and Aiding App™ are registered trademarks of
+      Canyon GBS LLC, and we are committed to enforcing and protecting our trademarks
+      vigorously.
+    - The software solution, including services, infrastructure, and code, is offered as a
+      Software as a Service (SaaS) by Canyon GBS LLC.
+    - Use of this software implies agreement to the license terms and conditions as stated
+      in the Elastic License 2.0.
+
+    For more information or inquiries please visit our website at
+    <https://www.canyongbs.com> or contact us via email at legal@canyongbs.com.
+
+</COPYRIGHT>
+*/
+
+namespace AidingApp\Contact\Tests\Tenant\Organization\RequestFactories;
+
+use AidingApp\Contact\Models\OrganizationIndustry;
+use AidingApp\Contact\Models\OrganizationType;
+use Worksome\RequestFactories\RequestFactory;
+
+class CreateOrganizationRequestFactory extends RequestFactory
+{
+    public function definition(): array
+    {
+        return [
+            'name' => fake()->company(),
+            'email' => fake()->companyEmail(),
+            'domains' => array_map(fn () => ['domain' => fake()->domainName()], range(1, 3)),
+            'phone_number' => fake()->phoneNumber(),
+            'website' => fake()->url(),
+            'industry_id' => OrganizationIndustry::factory(),
+            'type_id' => OrganizationType::factory(),
+            'description' => fake()->text(),
+            'number_of_employees' => fake()->randomNumber(),
+            'address' => fake()->address(),
+            'city' => fake()->city(),
+            'state' => fake()->state(),
+            'postalcode' => fake()->postcode(),
+            'country' => fake()->country(),
+            'linkedin_url' => fake()->url(),
+            'facebook_url' => fake()->url(),
+            'twitter_url' => fake()->url(),
+        ];
+    }
+}
